ihhashi/backend/requirements.txt
fastapi[standard]>=0.109.0
motor>=3.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
stripe>=7.10.0
twilio>=8.12.0
redis>=5.0.1
celery>=5.3.6
httpx>=0.26.0
supabase>=2.3.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=1.40.0
posthog>=3.5.0
python-telegram-bot>=21.0
bleach>=6.1.0
prometheus-client>=0.19.0
structlog>=24.1.0

# Route optimization
ortools>=9.8.0

# Quantum routing (D-Wave Leap)
dwave-system>=1.25.0
dwave-networkx>=0.8.12
dimod>=0.12.12

# Quantum routing (Qiskit QAOA)
qiskit>=1.0.0
qiskit-optimization>=0.6.0
qiskit-algorithms>=0.3.0
pyqubo>=1.4.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
mongomock>=4.1.2
